* Removes redundant code from include-text-cancelling-plugins onPaste method * Adds missing test for include-text-cancelling-plugins * Moves tests which ended up convering include-text-cancelling-plugins from Break/editor-registration.test.js into the include-text-cancelling-plugins test
